clean_requirements.txt
# Core dependencies for Subsplash Calendar Sync
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2

# Google Calendar API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Web scraping and browser automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# HTML parsing
lxml>=4.9.0

# Environment variable loading
python-dotenv>=1.0.0
